requirements.txt
flask
twilio
openai
chromadb
langchain
langchain-community
langchain-openai
python-dotenv